add auth shemas
