+++ models/contact.js
@@ -1,6 +1,9 @@
 const { Schema, model } = require("mongoose");
+const Joi = require("joi");
+
 const { handleMongooseError } = require("../helpers");
 
+// Mongoose shemas
 const contactSchema = new Schema(
   {
     name: {
@@ -27,4 +30,27 @@ contactSchema.post("save", handleMongooseError);
 
 const Contact = model("contact", contactSchema);
 
-module.exports = Contact;
+// Joi shemas
+const addShema = Joi.object({
+  name: Joi.string().required(),
+  email: Joi.string().email().required(),
+  phone: Joi.string().required(),
+  favorite: Joi.boolean(),
+});
+const changeShema = Joi.object({
+  name: Joi.string(),
+  email: Joi.string().email(),
+  phone: Joi.string(),
+});
+
+const upDateFavoriteSchema = Joi.object({
+  favorite: Joi.boolean().required(),
+});
+
+const schemas = {
+  addShema,
+  changeShema,
+  upDateFavoriteSchema,
+};
+
+module.exports = { Contact, schemas };
