+++ models/users.js
@@ -1,0 +1,25 @@
+const { Schema, model } = require("mongoose");
+const Joi = require("joi");
+const { handleMongooseError } = require("../helpers");
+
+const emailRegexp = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/;
+
+const userShema = new Shema(
+  {
+    name: {
+      type: String,
+      require: true,
+    },
+    email: {
+      type: String,
+      match: emailRegexp,
+      require: true,
+    },
+    password: {
+      type: String,
+      minlength: 6,
+      require: 6,
+    },
+  },
+  { versionKey: false }
+);
