+++ routes/api/auth.js
@@ -1,0 +1,7 @@
+const express = require("express");
+
+const router = express.Router();
+
+router.post("/register");
+
+module.exports = router;
