+++ routes/api/contacts.js
@@ -11,7 +11,7 @@ const {
   validatePatchBody,
 } = require("../../middlewares");
 
-const schemas = require("../../schemas/contacts");
+const { schemas } = require("../../models/contact");
 
 router.get("/", ctrl.getAllContacts);
