--- schemas/contacts.js
@@ -1,23 +1,0 @@
-const Joi = require("joi");
-
-const addShema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().email().required(),
-  phone: Joi.string().required(),
-  favorite: Joi.boolean(),
-});
-const changeShema = Joi.object({
-  name: Joi.string(),
-  email: Joi.string().email(),
-  phone: Joi.string(),
-});
-
-const upDateFavoriteSchema = Joi.object({
-  favorite: Joi.boolean().required(),
-});
-
-module.exports = {
-  addShema,
-  changeShema,
-  upDateFavoriteSchema,
-};
